% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patstater.r
\name{patstatr.filter}
\alias{patstatr.filter}
\title{Filter tables of PatStat raw data}
\usage{
patstatr.filter(file.pattern, ..., file.dir = file.path(getwd(),
  "patstat-rds"), progress.bar = TRUE, cols = character(0))
}
\arguments{
\item{file.pattern}{A pattern for getting a file or a set of files (data batches)}

\item{...}{A filtering conditions to fetch certain rows. (See dplyr::filter)}

\item{file.dir}{A path to directory with .rds files. Default is in working directory "patstat-rds".}

\item{progress.bar}{Whether to show progress bar (with pbapply package). Default is TRUE}

\item{cols}{Which column to select. Default is all columns.}
}
\value{
A data.table with a subset the data.
}
\description{
Similar to dplyr::filter but for tables of PatStat raw data data saved in multiple .rds files
}
