% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patstater.r
\name{patstatr.save.rds}
\alias{patstatr.save.rds}
\title{Reads PatStat .txt files and saves them to .rds (also optional in batches)}
\usage{
patstatr.save.rds(file, dir = file.path(getwd(), "patstat-txt"),
  dir.rds = file.path(getwd(), "patstat-rds"), batch.lines = 0,
  file.lines = 0)
}
\arguments{
\item{file}{File name (.txt is expected)}

\item{dir}{Directory where the file is. Default is in working directory "patstat-txt"}

\item{dir.rds}{Where to save. Default is in working directory "patstat-rds"}

\item{batch.lines}{How many lines is to read in one batch (10^7 is recomended). The default is 0, meanning that reading will be done in one batch as a single file.}

\item{file.lines}{Length of the .tsv file. The default is 0. If it is not changed and batch.lines is specified then it will try to calculate it with grep.}
}
\value{
Saved file(s) path.
}
\description{
Reads PatStat .txt files and saves them to .rds (also optional in batches)
}
